Feature/elements add icons (#7)
* Element icons class

* Added icons and section-header component

* Finished 51472: [UI] + [Functionality] View tree view

* Moved treeview to a new class
