+++ source/nodejs/adaptivecards-designer/src/app.ts
@@ -7,6 +7,7 @@ import { OutlookContainer } from "./containers/outlook-container";
 import { CortanaContainer } from "./containers/cortana-container";
 import { SkypeContainer } from "./containers/skype-container";
 import { adaptiveCardSchema } from "./adaptive-card-schema";
+import Treeview from "./components/treeview";
 
 declare var monacoEditor: any;
 declare function loadMonacoEditor(schema, callback);
@@ -131,7 +132,7 @@ class DesignerApp {
     private _card: Adaptive.AdaptiveCard;
     private _hostContainerPicker: Controls.DropDown;
     private _selectedHostContainer: HostContainer;
-    private _treeViewFoldedElements: Array<string>;
+    private _treeViewComponent: Treeview;
 
     public buildTreeViewSheet(peer: Designer.DesignerPeer) {
         if (this.treeViewSheetHostElement) {
@@ -139,102 +140,9 @@ class DesignerApp {
             treeview.innerHTML = "";
 
             const items = [...this._card.getItems(), ...this._card.getActions()];
-            const listItems = this.generateTreeViewElements(items, peer);
+            this._treeViewComponent.updateDesigner(this._designer);
+            const listItems = this._treeViewComponent.generateTreeViewElements(items, peer);
             treeview.appendChild(listItems);
-        }
-    }
-
-    private generateTreeViewElements(cardItems: Array<Adaptive.CardElement | Adaptive.Action>, peer: Designer.DesignerPeer, identationLevel: number = 1): HTMLElement {
-        if (!cardItems || cardItems.length === 0) {
-            let node = document.createElement("ul");
-            return node;
-        }
-
-        let selected_id: string;
-        if  (peer instanceof Designer.ActionPeer) {
-            selected_id = peer ? peer.action.elementId : "";
-        } else if(peer instanceof Designer.CardElementPeer) {
-            selected_id = peer ? peer.cardElement.elementId : "";
-        }
-
-        let itemList = document.createElement("ul");
-        itemList.className = "treeview__container";
-
-        let itemIndex = 0;
-        while(itemIndex < cardItems.length) {
-            let item = cardItems[itemIndex];
-            const listItem = this.createTreeViewListItem(item, selected_id, identationLevel);
-            itemList.appendChild(listItem);
-
-            const isFolded = this._treeViewFoldedElements.indexOf(item.elementId) !== -1;
-
-            if ([Adaptive.Container.name, Adaptive.Column.name].indexOf(item.getJsonTypeName()) !== -1) {
-                itemList.appendChild(this.createChildList((item as Adaptive.Container).getItems(), peer, identationLevel, isFolded));
-            } else if (item.getJsonTypeName() === Adaptive.ColumnSet.name) {
-                itemList.appendChild(this.createChildList((item as Adaptive.ColumnSet).getColumns(), peer, identationLevel, isFolded));
-            }
-
-            itemIndex++;
-        }
-        return itemList;
-    }
-
-    private createChildList(items, peer, identationLevel, isFolded): HTMLElement {
-        const newItem = document.createElement("li");
-        newItem.className = "treeview__element";
-        if (isFolded) {
-            newItem.className += " is-folded";
-        }
-        newItem.appendChild(this.generateTreeViewElements(items, peer, ++identationLevel));
-        return newItem;
-    }
-
-    private createTreeViewListItem (item: Adaptive.CardElement | Adaptive.Action, selected_id: string, identationLevel: number): HTMLElement {
-        let listItem = document.createElement("li");
-        listItem.className = "treeview__element";
-        if (selected_id && item.elementId === selected_id) {
-            listItem.className += " is-selected";
-        }
-
-        if ([Adaptive.Container.name, Adaptive.Column.name, Adaptive.ColumnSet.name].indexOf(item.getJsonTypeName()) !== -1) {
-            let foldArrow = document.createElement("button");
-            foldArrow.className = `btn treeview__icon treeview__icon--arrow`;
-            if (this._treeViewFoldedElements.indexOf(item.elementId) !== -1) {
-                foldArrow.classList.toggle("is-rotated");
-            }
-            foldArrow.addEventListener("click", () => {
-                this.foldTreeViewContainer(item.elementId, foldArrow);
-                foldArrow.classList.toggle("is-rotated");
-            });
-            listItem.appendChild(foldArrow);
-        }
-
-        let icon = document.createElement("span");
-        icon.className = `treeview__icon treeview__icon--${item.getJsonTypeName().toLowerCase()}`;
-        listItem.appendChild(icon);
-
-        let title = document.createElement("span");
-        title.className = `treeview__title`;
-        title.textContent = item.getJsonTypeName();
-        listItem.appendChild(title);
-
-        listItem.addEventListener("click", () => {
-            this._designer.setSelectedPeerById(item.elementId);
-        });
-        listItem.style.paddingLeft = `${identationLevel * 24}px`;
-        return listItem;
-    }
-
-    private foldTreeViewContainer(elementId: string, listElement: HTMLElement) {
-        const parent = listElement.parentElement;
-        const childList = parent.nextElementSibling as HTMLElement;
-        let addClassResult = childList.classList.toggle("is-folded")
-
-        if (addClassResult) {
-            this._treeViewFoldedElements.push(elementId);
-        } else {
-            let index = this._treeViewFoldedElements.indexOf(elementId);
-            this._treeViewFoldedElements.splice(index, 1);
         }
     }
 
@@ -383,7 +291,6 @@ class DesignerApp {
     commandListHostElement: HTMLElement;
 
     constructor(designerHostElement: HTMLElement) {
-        this._treeViewFoldedElements = [];
         this._designerHostElement = designerHostElement;
 
         this.addContainers();
@@ -391,6 +298,7 @@ class DesignerApp {
         this._selectedHostContainer = this.hostContainers[0];
 
         this.recreateDesigner();
+        this._treeViewComponent = new Treeview(this._designer);
     }
 
     createContainerPicker(): Controls.DropDown {
